+++ src/Infrastructure/Test/Repositories/InMemoryTaxonTreeRepository.php
@@ -36,35 +36,48 @@ final class InMemoryTaxonTreeRepository implements TaxonTreeRepository, Category
     {
         return TaxonTree::fromIterable($this->getTaxonNodes())
             ->sort('order')
-            ->eachRecursive(fn ($node) => $node->setLocale($this->locale));
+            ->eachRecursive(fn($node) => $node->setLocale($this->locale));
+    }
+
+    public function getTreeByTaxonomy(string $taxonomyId): TaxonTree
+    {
+        return TaxonTree::fromIterable($this->getTaxonNodes($taxonomyId))
+            ->sort('order')
+            ->eachRecursive(fn($node) => $node->setLocale($this->locale));
     }
 
     public function findTaxonById(string $id): TaxonNode
     {
-        return $this->getTree()->find(fn (TaxonNode $taxonNode) => $taxonNode->getId() == $id);
+        return $this->getTree()->find(fn(TaxonNode $taxonNode) => $taxonNode->getId() == $id);
     }
 
     public function findTaxonByKey(string $key): TaxonNode
     {
-        return $this->getTree()->find(fn (TaxonNode $taxonNode) => $taxonNode->getKey() == $key);
+        return $this->getTree()->find(fn(TaxonNode $taxonNode) => $taxonNode->getKey() == $key);
     }
 
-    private function getTaxonNodes(): TaxonNodes
+    private function getTaxonNodes(?string $taxonomyId = null): TaxonNodes
     {
         $nodes = [];
 
         $taxonNodeClass = $this->container->get(TaxonNode::class);
 
         foreach (InMemoryTaxonRepository::$taxons as $taxon) {
+
+            if ($taxonomyId && $taxon->taxonomyId->get() !== $taxonomyId) {
+                continue;
+            }
+
             $nodes[] = $taxonNodeClass::fromMappedData([
                 'taxon_id' => $taxon->taxonId->get(),
+                'taxonomy_id' => $taxon->taxonomyId->get(),
                 'parent_id' => $taxon->getMappedData()['parent_id'],
                 'data' => json_encode($taxon->getData()),
                 'state' => $taxon->getMappedData()['state'],
                 'order' => $taxon->getMappedData()['order'],
                 'product_ids' => $this->getCommaSeparatedProductIds($taxon->taxonId),
                 'online_product_ids' => $this->getCommaSeparatedOnlineProductIds($taxon->taxonId),
-                'keys' => json_encode(array_map(fn ($taxonKey) => $taxonKey->getMappedData(), $taxon->getTaxonKeys())),
+                'keys' => json_encode(array_map(fn($taxonKey) => $taxonKey->getMappedData(), $taxon->getTaxonKeys())),
             ]);
         }
 
@@ -73,7 +86,7 @@ final class InMemoryTaxonTreeRepository implements TaxonTreeRepository, Category
 
     private function getCommaSeparatedProductIds(TaxonId $taxonId): string
     {
-        if (! isset(InMemoryTaxonRepository::$productIds[$taxonId->get()])) {
+        if (!isset(InMemoryTaxonRepository::$productIds[$taxonId->get()])) {
             return '';
         }
 
@@ -82,7 +95,7 @@ final class InMemoryTaxonTreeRepository implements TaxonTreeRepository, Category
 
     private function getCommaSeparatedOnlineProductIds(TaxonId $taxonId): string
     {
-        if (! isset(InMemoryTaxonRepository::$onlineProductIds[$taxonId->get()])) {
+        if (!isset(InMemoryTaxonRepository::$onlineProductIds[$taxonId->get()])) {
             return '';
         }